requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.5.3
plotly==5.15.0
streamlit==1.29.0
